fix(ai): clear cumulative-delta CI max-lines and parity
Move stream helpers out of openai-completions-stream.ts and align the
intentional long refusal-repeat parity case with the >=8 exact-replay contract.
